我们的数组元素超越了int

+++ test-gpu.cu
@@ -4,7 +4,7 @@
  * @version: 
  * @Date: 2024-11-21 13:55:45
  * @LastEditors: Ruinique
- * @LastEditTime: 2025-01-02 15:24:32
+ * @LastEditTime: 2025-01-02 15:39:43
  */
 #include "common.h"
 #include "lu-factorization-with-pivoting-cuslover.h"
@@ -26,7 +26,7 @@ int main() {
     float time[14];
     size_t memory[14];
     for (int i = 0; i < 14; i++) {
-        int size = testSize[i];
+        size_t size = testSize[i];
         size_t memory_usage = calculate_memory_usage(size, size, size);
         memory[i] = memory_usage;
         float *matrix = new float[size * size];
